Seperate index.js into modules
