--- src/tabs/todo.js
@@ -1,7 +1,0 @@
-const toDoList = (title, list = []) => ({ title, list });
-
-const task = (title, desc, dueDate, priority, status = false) => ({
-  title, desc, dueDate, priority, status,
-});
-
-export { toDoList, task };
